Resolve ambiguous call to sqrt.
git-svn-id: svn://svn.code.sf.net/p/stxxl/code/trunk@941 4e380d45-d1fd-0310-85a7-d18ec86df0ad

+++ utils/pq_param.cpp
@@ -49,8 +49,8 @@ int main()
                 cout << "c: " << c << endl;
                 cout << "k: " << k << endl;
                 int64 Ae = m / 2;
-                int64 Ae1 = Ae + int64(sqrt(D(m, k, MaxS, E, B)));
-                int64 Ae2 = Ae - int64(sqrt(D(m, k, MaxS, E, B)));
+                int64 Ae1 = Ae + int64(sqrt((double)D(m, k, MaxS, E, B)));
+                int64 Ae2 = Ae - int64(sqrt((double)D(m, k, MaxS, E, B)));
                 int64 x = c * B / E;
                 int64 N = x / 4096;
                 cout << "Ae : " << Ae << endl;
